(Atualizando)

+++ PROJETOUNICO/src/controlador/ExcluiControlador.java
@@ -11,15 +11,11 @@ public class ExcluiControlador implements TemplateViewRoute {
 	
 	private LivroDAO dao = new LivroDAO();
 
-	@Override
 	public ModelAndView handle(Request req, Response resp) {		
 		int numero = Integer.parseInt(req.params("ISBN")); // /excluir/:numero
 		Livro livro = dao.load(numero);
-		if (livro != null) {
-			dao.delete(livro);
-		}		
+		if (livro != null) dao.delete(livro);
 		resp.redirect("/lista");
 		return null;
 	}
-
 }
